requirements.txt
# Core ML libraries
keras==2.13.1
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
Pillow==10.0.0
scikit-learn==1.3.0

# Image processing
opencv-python==4.8.0.76
imageio==2.31.1

# Web framework and API
flask==2.3.2
flask-cors==4.0.0
werkzeug==2.3.6

# File handling
pathlib2==2.3.7

# Load testing
locust==2.15.1

# Visualization and monitoring
plotly==5.15.0
dash==2.11.1

# Utilities
tqdm==4.65.0
joblib==1.3.1